Expose execute_with_pipes and the needed functions. To make writing tests in depending applications easier.
